Excluded range added with user range. You can add and calculate with your own filter

+++ RF_Go/Models/RFChannels.cs
@@ -56,7 +56,7 @@ namespace RF_Go.Models
                 }
             }
         }
-        public void SetRandomFrequency(HashSet<int> UsedFrequencies, HashSet<int> TwoTX3rdOrder, HashSet<int> TwoTX5rdOrder, HashSet<int> TwoTX7rdOrder, HashSet<int> TwoTX9rdOrder, HashSet<int> ThreeTX3rdOrder, List<(int StartFrequency, int EndFrequency)> excludedRanges)
+        public void SetRandomFrequency(HashSet<int> UsedFrequencies, HashSet<int> TwoTX3rdOrder, HashSet<int> TwoTX5rdOrder, HashSet<int> TwoTX7rdOrder, HashSet<int> TwoTX9rdOrder, HashSet<int> ThreeTX3rdOrder, List<(float StartFrequency, float EndFrequency)> excludedRanges)
         {
             int f1;
                 if (!Checked && !IsLocked)
@@ -109,7 +109,7 @@ namespace RF_Go.Models
             int randomIndex = random.Next(0, numberOfValue);
             return (Range[0] + (randomIndex * Range[3]));
         }
-        public int LoopForFreeFrequency(HashSet<int> UsedFrequencies, HashSet<int> TwoTX3rdOrder, HashSet<int> TwoTX5rdOrder, HashSet<int> TwoTX7rdOrder, HashSet<int> TwoTX9rdOrder, HashSet<int> ThreeTX3rdOrder, List<(int StartFrequency, int EndFrequency)> excludedRanges)
+        public int LoopForFreeFrequency(HashSet<int> UsedFrequencies, HashSet<int> TwoTX3rdOrder, HashSet<int> TwoTX5rdOrder, HashSet<int> TwoTX7rdOrder, HashSet<int> TwoTX9rdOrder, HashSet<int> ThreeTX3rdOrder, List<(float StartFrequency, float EndFrequency)> excludedRanges)
         {
             int f1;
             int count = 0;
@@ -135,7 +135,7 @@ namespace RF_Go.Models
 
             return 0;
         }
-        public bool CheckFreeFrequency(int f1, HashSet<int> UsedFrequencies, HashSet<int> TwoTX3rdOrder, HashSet<int> TwoTX5rdOrder, HashSet<int> TwoTX7rdOrder, HashSet<int> TwoTX9rdOrder, HashSet<int> ThreeTX3rdOrder, List<(int StartFrequency, int EndFrequency)> excludedRanges)
+        public bool CheckFreeFrequency(int f1, HashSet<int> UsedFrequencies, HashSet<int> TwoTX3rdOrder, HashSet<int> TwoTX5rdOrder, HashSet<int> TwoTX7rdOrder, HashSet<int> TwoTX9rdOrder, HashSet<int> ThreeTX3rdOrder, List<(float StartFrequency, float EndFrequency)> excludedRanges)
         {
             // Check if the frequency falls within any of the exclusion ranges
